// apps/wasm4pm/src/commands/drift-watch.d.ts
export declare const driftWatch: import('citty').CommandDef<{
  input: {
    type: 'string';
    description: string;
    required: true;
    alias: string;
  };
  'activity-key': {
    type: 'string';
    description: string;
    alias: string;
  };
  window: {
    type: 'string';
    description: string;
    alias: string;
  };
  interval: {
    type: 'string';
    description: string;
    alias: string;
  };
  alpha: {
    type: 'string';
    description: string;
  };
  threshold: {
    type: 'string';
    description: string;
  };
  json: {
    type: 'boolean';
    description: string;
  };
  enhanced: {
    type: 'boolean';
    description: string;
  };
}>;
//# sourceMappingURL=drift-watch.d.ts.map